added a check script and corrected copy script. also added a bash examples/testing script

+++ copy_files.sh
@@ -2,23 +2,23 @@
 #script designed to perform an md5sum check of files being copied
 
 
-rm /movies/md5check.txt
-rm /movies/md5checkreceiving.txt
-rm /movies/checkincoming.txt
-rm /movies/checksource.txt
+
+rm -f /tmp/md5check_source.txt /tmp/md5check_dest.txt
+read -p "Are files BluRay (yes or no): " bluray
+echo $bluray
 for each in `cat ~/files.to.copy`
 do
-cd /movies/
-md5sum /costamovies/$each >> /movies/md5check.txt
-md5sum --check md5check_incoming.txt
-
+echo $each
+if [ "$bluray" = no ]
+then
 cp /costamovies/$each /movies
-
-md5sum /movies/$each >> /movies/md5checkreceiving.txt
-md5sum --check md5checkreceiving.txt
-
-cat /movies/md5checkreceiving.txt | cut -d" " -f1 > /movies/check_destination.txt
-cat /movies/md5check_source.txt | cut -d" " -f1 > /movies/checksource.txt
+#md5sum /costamovies/$each >> /tmp/md5check_source.txt
+#md5sum /movies/$each >> /tmp/md5check_dest.txt
+else
+cp /costamovies/BluRay/$each /movies/BluRay/
+#md5sum /costamovies/BluRay/$each >> /tmp/BluRay/md5check_source.txt
+#md5sum /movies/BluRay/$each >> /tmp/BluRay/md5check_dest.txt
+fi
 
 done
 
@@ -26,6 +26,6 @@ done
 #sed -r 's/^(\S+\s+){2}//' /movies/md5checkreceiving.txt
 #sed -r 's/^(\S+\s+){2}//' /movies/md5check.txt
 
-diff /movies/checksource.txt /movies/check_destination.txt
+#diff /movies/checksource.txt /movies/check_destination.txt
 
 exit 0
